requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
Django==3.2.18
djangorestframework==3.14.0
drf-spectacular==0.25.1
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
inflection==0.5.1
jsonschema==4.17.3
pkgutil-resolve-name==1.3.10
psycopg2-binary==2.8.5
pyrsistent==0.19.3
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
sqlparse==0.4.3
typing-extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.14
zipp==3.15.0
